Splitted test into unit / integration tests
